Edits based on comments PR
