--- Library/Homebrew/os/linux.rb
@@ -1,10 +1,0 @@
-require 'hardware'
-
-module OS
-  module Linux
-    extend self
-
-    ::Linux = self # compatibility
-
-  end
-end
